fix(emby): address review feedback on the stale record cleanup
- Include the parent series id in the stale episode identity so a row
  that Emby re-linked to a different series is treated as stale instead
  of seen, and cleaned up rather than staying attached to the old series.
- Perform the multi-episode fill span subtraction as long in the episode
  and played syncs so a bogus negative IndexNumber cannot overflow the
  span and bypass the 50-episode cap.
- Clear the run-scoped seen id set at the start of the content sync in
  case the job instance is ever reused.
- Cover the played movie empty-page guard and the overflow guards with
  tests.

+++ src/Ombi.Schedule/Jobs/Emby/EmbyPlayedSync.cs
@@ -222,7 +222,9 @@ namespace Ombi.Schedule.Jobs.Emby
             // exist, or to run for hundreds of thousands of iterations.
             if (episode.IndexNumberEnd.HasValue && episode.IndexNumberEnd.Value > episode.IndexNumber)
             {
-                var episodeFillCount = episode.IndexNumberEnd.Value - episode.IndexNumber;
+                // Subtract as long so a bogus negative IndexNumber cannot overflow the
+                // span, bypass the cap below and drive a huge fill loop
+                var episodeFillCount = (long)episode.IndexNumberEnd.Value - episode.IndexNumber;
 
                 if (episodeFillCount > MaxEpisodeFillCount)
                 {
